chore: break noz peek hook into single file hooks, drop pb-28 from main footer
